# Optimised /api/v1/series for blocks storage