**What this PR does**:
While working on #2794, I've compared our implementation to fetch series with Prometheus one and I've noticed Prometheus is easier. I've done the refactoring and written and benchmark and turned out it performs better too.

```
benchmark                                           old ns/op     new ns/op     delta
Benchmark_Ingester_v2MetricsForLabelMatchers-12     250057594     226011946     -9.62%

benchmark                                           old allocs     new allocs     delta
Benchmark_Ingester_v2MetricsForLabelMatchers-12     1104071        800196         -27.52%

benchmark                                           old bytes     new bytes     delta
Benchmark_Ingester_v2MetricsForLabelMatchers-12     62724828      51382654      -18.08%
```

**Which issue(s) this PR fixes**:
N/A

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`